Fix firm for linux

+++ STM32/Core/Src/firm_tasks.c
@@ -9,6 +9,7 @@
 #include "usb_print_debug.h"
 #include "usbd_cdc_if.h"
 #include "semphr.h"
+#include <stdint.h>
 #include <string.h>
 
 // task handles
@@ -565,7 +566,8 @@ void transmit_data(void *argument) {
       packet.crc = crc16_ccitt((uint8_t *)&(packet), serialized_packet_len - 2);
       // move the location of the crc bytes directly after the payload
       memmove((uint8_t*)&packet + serialized_packet_len - 2, &packet.crc, sizeof(packet.crc));
-      CDC_Transmit_FS((uint8_t*)&packet, serialized_packet_len);
+      CDC_Transmit_FS((uint8_t*)&packet, (uint16_t)serialized_packet_len);
+
       // optionally transmit over uart if the setting is enabled
       if (firmSettings.uart_transfer_enabled && uart_tx_done) {
         uart_tx_done = false;
